-- [Read all these notes]: End of day push, should build ---- Fleshed out basis for instructor result; need to relook at requirements initial permise did not match my what I remember. Try to start down that road! ---- Tweeked the spresult model to support a composite scoring system as the "averaging" model revealed unconsidered flaws. ---- Cleanup the faculty publish controller and tried to elimate some redundacys.

+++ Ecat.Shared.Core/ModelLibrary/Learner/SpResponse.cs
@@ -22,7 +22,7 @@ namespace Ecat.Shared.Core.ModelLibrary.Learner
         public int InventoryItemId { get; set; }
 
         public string MpItemResponse { get; set; }
-        public float ItemModelScore { get; set; }
+        public int ItemModelScore { get; set; }
 
         public SpInventory InventoryItem { get; set; }
         public WorkGroup WorkGroup { get; set; }
